requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
anyio==4.0.0
APScheduler==3.6.3
asgiref==3.6.0
asttokens==2.4.1
async-timeout==4.0.2
atpublic==4.0
attrs==22.2.0
backcall==0.2.0
bcrypt==4.1.1
beautifulsoup4==4.12.2
bidict==0.22.1
bigframes==0.11.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2022.12.7
cffi==1.16.0
charset-normalizer==3.0.1
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudevents==1.9.0
cloudpickle==3.0.0
colorama==0.4.6
comm==0.1.4
cryptography==41.0.5
DateTime==5.4
db-dtypes==1.1.1
dbt==1.0.0.34.9
dbt-init==0.2.7
decorator==5.1.1
deprecation==2.1.0
Django==4.1.7
dnspython==2.4.2
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==2.0.0
filelock==3.12.4
fiona==1.9.5
firebase-admin==6.2.0
firebase-functions==0.1.2
Flask==2.3.3
Flask-Cors==4.0.0
frozenlist==1.3.3
fsspec==2023.10.0
functions-framework==3.4.0
gcsfs==2023.10.0
gdown==4.7.1
geopandas==0.14.0
google==3.0.0
google-api-core==2.12.0
google-api-python-client==2.113.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-bigquery==3.12.0
google-cloud-bigquery-connection==1.13.2
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-functions==1.13.3
google-cloud-iam==2.12.2
google-cloud-resource-manager==1.10.4
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-events==0.11.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.1
grpc-google-iam-v1==0.12.6
grpcio==1.59.0
grpcio-status==1.59.0
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httpx==0.25.0
humanize==4.8.0
ibis-framework==6.2.0
idna==3.4
ipython==8.16.1
ipywidgets==8.1.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyterlab-widgets==3.0.9
llvmlite==0.41.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
multipledispatch==1.0.0
numba==0.58.1
numpy==1.26.1
oauth2client==4.1.3
oauthlib==3.2.2
openai==0.27.0
openpyxl==3.1.2
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.1
parso==0.8.3
parsy==2.1
PGPy==0.6.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.24.4
proton==0.9.1
psutil==5.9.6
pure-eval==0.2.2
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybase64==1.3.1
pycparser==2.21
pydata-google-auth==1.8.2
PyDrive==1.3.1
Pygments==2.16.1
PyJWT==2.8.0
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyPDF2==3.0.1
pyproj==3.6.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-gnupg==0.5.2
python-telegram-bot==13.3
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.6.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.3
screeninfo==0.8.1
selenium==4.16.0
shapely==2.0.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.22
sqlglot==17.16.2
sqlparse==0.4.3
stack-data==0.6.3
threadpoolctl==3.2.0
toolz==0.12.0
tornado==6.3.3
tqdm==4.64.1
traitlets==5.12.0
trio==0.23.2
trio-websocket==0.11.1
typing_extensions==4.8.0
tzdata==2022.7
tzlocal==5.1
uritemplate==4.1.1
urllib3==1.26.14
watchdog==3.0.0
wcwidth==0.2.8
webdriver-manager==4.0.1
Werkzeug==3.0.1
widgetsnbextension==4.0.9
wsproto==1.2.0
xxhash==3.4.1
yarl==1.8.2
zope.interface==6.1